Improve handling of macros and parse failures
