+++ src/main/scala/viper/silver/cfg/CfgTest.scala
@@ -8,7 +8,6 @@ package viper.silver.cfg
 
 import java.nio.file.{Files, Path, Paths}
 
-import fastparse._
 import viper.silver.parser.{FastParser, PProgram, Resolver, Translator}
 import viper.silver.verifier.ParseWarning
 
@@ -36,12 +35,8 @@ object CfgTest {
   }
 
   private def parse(input: String, file: Path): Option[PProgram] = {
-    val result = new FastParser().parse(input, file)
-    result match {
-      case Parsed.Success(program@PProgram(_, _, _, _, _, _, _,_, errors), _) =>
-        if (errors.isEmpty || errors.forall(_.isInstanceOf[ParseWarning])) Some(program)
-        else None
-      case _ => None
-    }
+    val program = new FastParser().parse(input, file)
+    if (program.errors.forall(_.isInstanceOf[ParseWarning])) Some(program)
+    else None
   }
 }
