+++ src/test/resources/all/issues/silver/0138.vpr
@@ -2,8 +2,8 @@
 // http://creativecommons.org/publicdomain/zero/1.0/
 
 define outer
-//:: ExpectedOutput(parser.error)
 {
+//:: ExpectedOutput(parser.error)
   define inner {
   }
 }
